+++ lib/slide_button.dart
@@ -17,8 +17,8 @@ class SlideButton extends StatefulWidget {
   final double slideButtonIconSize;
   final double radius;
   final double successfulThreshold;
-  final Widget widgetWhenDragIsSuccess;
-  final VoidCallback onDragSuccessCallback;
+  final Widget widgetWhenSlideIsCompleted;
+  final VoidCallback onSlideSuccessCallback;
 
   const SlideButton({Key key,
     this.buttonHeight = 55,
@@ -32,24 +32,24 @@ class SlideButton extends StatefulWidget {
     this.slideButtonIconSize = 30.0,
     this.radius = 4.0,
     this.successfulThreshold = 0.9,
-    this.widgetWhenDragIsSuccess,
-    this.onDragSuccessCallback,
+    this.widgetWhenSlideIsCompleted,
+    this.onSlideSuccessCallback,
   }) : super(key: key);
 
   @override
-  State<StatefulWidget> createState() => _SlideButtonState(
+  State<StatefulWidget> createState() => SlideButtonState(
       defaultButtonSize: this.buttonHeight, defaultButtonColor: this.buttonColor,
       defaultButtonText: this.buttonText, defaultSlideButtonMargin: this.slideButtonMargin,
       defaultSlideButtonColor: this.slideButtonColor, defaultButtonTextColor: this.buttonTextColor,
       defaultSlideButtonIconColor: this.slideButtonIconColor, defaultSlideButtonIcon: this.slideButtonIcon,
       defaultSlideButtonIconSize: this.slideButtonIconSize, defaultRadius: this.radius,
-      successfulThreshold: this.successfulThreshold, widgetWhenDragIsSuccess: this.widgetWhenDragIsSuccess,
-      onDragSuccessCallback: this.onDragSuccessCallback
+      successfulThreshold: this.successfulThreshold, widgetWhenSlideIsCompleted: this.widgetWhenSlideIsCompleted,
+      onSlideSuccessCallback: this.onSlideSuccessCallback
   );
 
 }
 
-class _SlideButtonState extends State<SlideButton> {
+class SlideButtonState extends State<SlideButton> {
 
   final _buttonKey = GlobalKey();
   final _slideButtonKey = GlobalKey();
@@ -65,8 +65,8 @@ class _SlideButtonState extends State<SlideButton> {
   double defaultSlideButtonIconSize;
   double defaultRadius;
   double successfulThreshold;
-  Widget widgetWhenDragIsSuccess;
-  VoidCallback onDragSuccessCallback;
+  Widget widgetWhenSlideIsCompleted;
+  VoidCallback onSlideSuccessCallback;
 
   bool _isSlideEnabled = false;
   bool _isSlideStarted = false;
@@ -75,7 +75,7 @@ class _SlideButtonState extends State<SlideButton> {
   double _slideButtonSize;
   double _slideButtonMargin;
 
-  _SlideButtonState({
+  SlideButtonState({
     this.defaultButtonSize,
     this.defaultButtonText,
     this.defaultSlideButtonMargin,
@@ -87,8 +87,8 @@ class _SlideButtonState extends State<SlideButton> {
     this.defaultSlideButtonIconSize,
     this.defaultRadius,
     this.successfulThreshold,
-    this.widgetWhenDragIsSuccess,
-    this.onDragSuccessCallback,
+    this.widgetWhenSlideIsCompleted,
+    this.onSlideSuccessCallback,
   });
 
   @override
@@ -97,9 +97,9 @@ class _SlideButtonState extends State<SlideButton> {
     // Initialize properties used on the slide button
     _slideButtonSize = defaultButtonSize - (defaultSlideButtonMargin * 2);
     _slideButtonMargin = defaultSlideButtonMargin;
-    // Always add a default widget for drag successful event
-    if (this.widgetWhenDragIsSuccess == null) {
-      this.widgetWhenDragIsSuccess = Center(
+    // Always add a default widget for slide successful event
+    if (this.widgetWhenSlideIsCompleted == null) {
+      this.widgetWhenSlideIsCompleted = Center(
         child: SizedBox(
           width: defaultButtonSize / 3, height: defaultButtonSize / 3,
           child: PlatformProgressIndicator(materialValueColor: AlwaysStoppedAnimation<Color>(this.defaultSlideButtonIconColor), materialStrokeWidth: 1.3,),
@@ -117,10 +117,11 @@ class _SlideButtonState extends State<SlideButton> {
           // Check if the tap down event has occurred inside the slide button
           final RenderBox renderBox = _slideButtonKey.currentContext.findRenderObject();
           final slideButtonOffset = renderBox.localToGlobal(Offset.zero);
-          final startXPosition = slideButtonOffset.dx;
-          final endXPosition = slideButtonOffset.dx + _slideButtonSize;
-          final startYPosition = slideButtonOffset.dy;
-          final endYPosition = slideButtonOffset.dy + _slideButtonSize;
+          // On all positions I've added the _slideButtonMargin. Basically we use the _slideButtonMargin as a invisible touchable area that triggers the slide event
+          final startXPosition = slideButtonOffset.dx - _slideButtonMargin;
+          final endXPosition = startXPosition + defaultButtonSize + _slideButtonMargin;
+          final startYPosition = slideButtonOffset.dy - _slideButtonMargin;
+          final endYPosition = startYPosition + defaultButtonSize + _slideButtonMargin;
           // print(startXPosition);
           // print(endXPosition);
           // print(startYPosition);
@@ -183,7 +184,7 @@ class _SlideButtonState extends State<SlideButton> {
         onHorizontalDragEnd: (dragDetails) {
           print('horizontal_drag_end');
           if (_isSlideEnabled || _isSlideStarted) {
-            // Check if the drag event has reached the minimum threshold to be considered a successful drag event
+            // Check if the slide event has reached the minimum threshold to be considered a successful slide event
             final RenderBox renderBox = _buttonKey.currentContext.findRenderObject();
             if (_slideButtonSize >= successfulThreshold * renderBox.size.width) {
               _slideButtonSize = renderBox.size.width;
@@ -191,7 +192,7 @@ class _SlideButtonState extends State<SlideButton> {
               _isSlideEnabled = false;
               _isSlideStarted = false;
               // Make sure that we've called the success callback
-              onDragSuccessCallback?.call();
+              onSlideSuccessCallback?.call();
             } else {
               _slideButtonMarginDragOffset = 0;
               _slideButtonSize = defaultButtonSize - (defaultSlideButtonMargin * 2);
@@ -221,11 +222,18 @@ class _SlideButtonState extends State<SlideButton> {
                 AnimatedContainer(
                   key: _slideButtonKey,
                   margin: EdgeInsets.only(left: _slideButtonMargin, top: _slideButtonMargin),
-                  duration: Duration(milliseconds: 50),
+                  duration: Duration(milliseconds: 100),
                   width: _slideButtonSize, height: _slideButtonSize,
                   decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(this.defaultRadius),
                     color: defaultSlideButtonColor,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.black12,
+                        blurRadius: 5.0,
+                        spreadRadius: 2.0,
+                      )
+                    ],
                   ),
                   child: Center(
                     child: Icon(
@@ -240,7 +248,7 @@ class _SlideButtonState extends State<SlideButton> {
                     width: double.infinity, height: defaultButtonSize,
                     color: defaultSlideButtonColor,
                     child: Center(
-                      child: widgetWhenDragIsSuccess,
+                      child: widgetWhenSlideIsCompleted,
                     ),
                   ),
                 ),
@@ -250,6 +258,16 @@ class _SlideButtonState extends State<SlideButton> {
         ),
       ),
     );
+  }
+
+  void reset() {
+    _slideButtonMarginDragOffset = 0;
+    _slideButtonSize = defaultButtonSize - (defaultSlideButtonMargin * 2);
+    _slideButtonMargin = defaultSlideButtonMargin;
+    _hasCompletedSlideWithSuccess = false;
+    _isSlideEnabled = false;
+    _isSlideStarted = false;
+    setState(() {});
   }
 
 }