+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - First release with Slide Button 100% functional (maybe with few bugs)
 
-* TODO: Describe initial release.
+* First release with Slide Button 100% functional (maybe with few bugs). For more info check the README.md